+++ src/models/Inventory.Model.ts
