+++ src/routes/inventoryRoutes.ts
